pyproject.toml
[tool.poetry]
name = "ticket-office"
version = "0.1.0"
description = ""
authors = ["Saar Wasserman <saarwasserman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.1.5"
djangorestframework = "3.12.2"
requests = "^2.25.1"
djoser = "^2.1.0"
django-model-utils = "^4.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-django = "^4.1.0"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
